Added primary key and modes
Necessary to define the primary key used for tables. This constant will be used to determine default behavior when peforming find, create, update, and delete operations for example. The ability to change default behavior will also be supported.

The modes map to the following:

* `default`: Uses PStore default settings.
* `thread`: Enables thread safetey which has a slight performance hit but ensures all transactions ae thread safe.
* `file`: This is also known has `ultra` in the original `PStore` gem which makes a copy of the persinstant store before modifying it in case there are issues with the file system or operating system.
* `max`: Enables both of the above for maximum security.

These modes will soon be used when configuring the gem.