Added persistent store refinements
Necessary to refine and make native `PStore` functionality reasonable to work with since the original design is so poor especially in regards to boolean control coupling, lack of attributes, and a messy implementation in general.

Reek had to be updated because of the above issues as well. A necessary evil, sadly.