Added abstract table
Necessary to implement basic table functionality which can be specialized, via subclasses, for dictionary (hash), value objects, and other kinds of record types.

Reek was updated to ignore data clumps since being able to supply a custom key provides a useful way to use different keys for the same object. As for unused parameters, those are necessary for the subclass method signatures.