Added configuration
Necessary to configure this gem for use with PStore and to alleviate setup burden. A special register method was added to register common models and their associated primary keys for future use in tables. Finally, a `Struct` was chosen over `Data` because mutability will allow the client (not implemented yet) to use a block for configuration during initalization.

RuboCop was updated to ignore long parameter lists because RuboCop doesn't fully recognize the value of a Struct object.

Reek was updated to ignore module initialize since it doesn't know how to handle Struct objects yet.