Added table record creation
Necessary to expose basic record creation if upserting is not desired. The record is only created if not found.

The protected `#primary_id` method was introduced so each subclass could implement accordingly while allowing the abstract class to handle the implementation.

Reek was updated to ignore the `#primary_id` violation for unused parameters since this is meant to be implemented by the subclasses.

Milestone: minor